scripts: start.sh: add --network host to docker options
Add `--network host` to the Docker run options in `scripts/start.sh`
to allow the container to use the host network stack. This is useful
for networking tests, simplifying port bindings, and ensuring direct
network access in sandboxed environments.

The change retains backward compatibility with other existing Docker
flags used in the script.

Signed-off-by: Miguel Miranda <mmikernel@gmail.com>